Extract three self-contained subsystems out of ParakeetEngine.swift
ParakeetEngine.swift had grown back to 3813 lines after the 2026-07
splits because three later subsystems were grown in-place instead of
following the established extension-file pattern (2026-08-05 post-program
audit finding). Move them, no behavior change:

- ParakeetSystemInputCoordination.swift — system default-input restore,
  owner-bound pending-restore handoff, and late-completion reconciliation
- ParakeetSharedMeetingMicBridge.swift — borrowed-mic dictation bridge
  (start/finish, PCM intake, resume-after-meeting, stale-claim finalize)
- ParakeetZombieEngineRecovery.swift — startup audio watchdog plus the
  generation-gated zombie-engine replacement task

ParakeetEngine.swift drops to 3025 lines and stays the public-API owner
and @MainActor home for all moved state; members the extensions touch
widen from private to internal, matching ParakeetDeviceRecovery.swift.
The audio-graph/queue owner-token helpers stay in ParakeetEngine.swift
because every start/stop path uses them, not just recovery.

Source-contract tests that pinned the moved code to ParakeetEngine.swift
now read the new files (new readParakeetZombieRecoverySource /
readParakeetSystemInputSource helpers); anchors tracking the widened
access levels updated in place. Sources/Speech/CLAUDE.md documents the
three new extension files.

Verified: swiftc typecheck, bash build.sh --no-open (with refreshed
deps), bash run-tests.sh (12774 passed).
