+++ Sources/Speech/CLAUDE.md
@@ -6,9 +6,12 @@
 
 ## Key Files
 
-- `ParakeetEngine.swift` — app-owned Parakeet STT engine, recording control, final dictation transcription, permission-aware input-readiness checks, short-audio gating, live level metering, and sanitized failure reporting for model init errors. Device-change recovery and model load/download/warmup/teardown are split into the two files below — ParakeetEngine remains the public-API owner and `@MainActor` home for that state; the split files are internal collaborator extensions.
+- `ParakeetEngine.swift` — app-owned Parakeet STT engine, recording control, final dictation transcription, permission-aware input-readiness checks, short-audio gating, live level metering, and sanitized failure reporting for model init errors. Device-change recovery, model lifecycle, system-input restore/reconciliation, zombie-engine recovery, and shared-meeting-mic bridging are split into the extension files below — ParakeetEngine remains the public-API owner and `@MainActor` home for that state; the split files are internal collaborator extensions.
 - `ParakeetDeviceRecovery.swift` — `ParakeetEngine` extension: device-change detection (CoreAudio default-input listener, `AVAudioEngineConfigurationChange` observer) and the `attemptDeviceRecovery` / `scheduleConfigRecoveryTimeout` executor that rewarms the audio graph after a route change. The pure decision tables it consults (`ParakeetDeviceRecoveryReadinessPolicy`, `ParakeetDeviceRecoveryFailurePolicy`, `ParakeetDeviceRecoveryTimeoutPolicy`) live in `ParakeetStartRecordingFailurePolicy.swift`, not here.
 - `ParakeetModelLifecycle.swift` — `ParakeetEngine` extension: model load/download/warmup/teardown paths (`initialize`, `performInitialize`, `prefetchModelFilesIfNeeded`, `cancelModelWork`, `teardownModel`)
+- `ParakeetSystemInputCoordination.swift` — `ParakeetEngine` extension: restoring the user's previous system default input after dictation temporarily overrode it (faster-Bluetooth-dictation path) — the owner-bound pending-restore handoff, bounded-timeout CoreAudio writes through the replaceable system-input work coordinator, and the reconciliation queue that converges late (timed-out) HAL completions back onto current MainActor intent
+- `ParakeetSharedMeetingMicBridge.swift` — `ParakeetEngine` extension: borrowed-mic dictation while a meeting owns the microphone — start/finish of borrowed-mic recording, relayed PCM buffer and level intake, resuming regular mic capture when the meeting ends first, and stale-claim resolution/finalization (see `SharedMeetingMicClaim.swift` for the handshake rationale)
+- `ParakeetZombieEngineRecovery.swift` — `ParakeetEngine` extension: the startup audio watchdog that detects zombie engines (running but no usable signal after sleep/wake) and the separate generation-gated recovery task that replaces the stale `AVAudioEngine` through a bounded reset and retries once; the audio-graph/queue owner-token helpers stay in `ParakeetEngine.swift` because every start/stop path uses them
 - `ParakeetAudioDeviceLookup.swift` — CoreAudio default-input lookup and dictation input selection descriptors used before Parakeet starts recording
 - `ParakeetAudioEngineSupport.swift` — support types for Parakeet engine startup snapshots, retired-engine retention, and default-input listener teardown
 - `WhisperEngine.swift` — app-owned WhisperKit STT engine used when advanced users select a Whisper model
